src/data_drift_backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python==8.0.31
catboost
eurybia
scikit-learn==1.1.1
joblib==1.1.0
pandas==1.3.3
numpy==1.21.2